+++ src/com/maddyhome/idea/vim/common/Marks.kt
@@ -1,0 +1,57 @@
+package com.maddyhome.idea.vim.common
+
+import com.intellij.ide.bookmarks.Bookmark
+import com.intellij.ide.bookmarks.BookmarkManager
+import com.intellij.openapi.project.Project
+import com.maddyhome.idea.vim.group.MarkGroup
+import java.lang.ref.WeakReference
+
+interface Mark {
+  val key: Char
+  val logicalLine: Int
+  val col: Int
+  val filename: String
+  val protocol: String
+
+  fun isClear(): Boolean
+  fun clear()
+}
+
+data class VimMark(
+  override val key: Char,
+  override var logicalLine: Int,
+  override val col: Int,
+  override val filename: String,
+  override val protocol: String
+) : Mark {
+
+  private var cleared = false
+
+  override fun isClear(): Boolean = cleared
+
+  override fun clear() {
+    cleared = true
+  }
+}
+
+class IntellijMark(bookmark: Bookmark, override val col: Int, project: Project?) : Mark {
+
+  val project: WeakReference<Project?> = WeakReference(project)
+
+  override val key = bookmark.mnemonic
+  override val logicalLine: Int
+    get() = getMark()?.line ?: 0
+  override val filename: String
+    get() = getMark()?.file?.path ?: ""
+  override val protocol: String
+    get() = getMark()?.file?.let { MarkGroup.extractProtocol(it) } ?: ""
+
+  override fun isClear(): Boolean = getMark()?.isValid?.not() ?: false
+  override fun clear() {
+    val mark = getMark() ?: return
+    project.get()?.let { project -> BookmarkManager.getInstance(project).removeBookmark(mark) }
+  }
+
+  private fun getMark(): Bookmark? =
+    project.get()?.let { project -> BookmarkManager.getInstance(project).findBookmarkForMnemonic(key) }
+}