+++ CHANGES.md
@@ -50,6 +50,7 @@ _To Be Released:_
 * [VIM-1804](https://youtrack.jetbrains.com/issue/VIM-1804) Exit insert move after toggling IdeaVim
 * [VIM-1749](https://youtrack.jetbrains.com/issue/VIM-1749) Tag surround is finished on `>`
 * [VIM-1801](https://youtrack.jetbrains.com/issue/VIM-1801) Fix j/k motions with inline hints
+* [VIM-1800](https://youtrack.jetbrains.com/issue/VIM-1800) Improve ideamarks option
 
 0.53, 2019-08-07
 --------------